requirements.txt
openpyxl
tqdm
Pillow
imagehash
